requirements.txt
pyOpenSSL == 21.0.0
urllib3==1.26.7
certifi==2021.10.8
requests==2.26.0
cryptography==36.0.1